Move loggers to their own file
