+++ game_interface.rb
@@ -1,44 +1,6 @@
 require 'concurrent'
 require "./direction.rb"
 
-class BaseLogger
-
-  def print_permanents(player)
-    prompt="here are the permanents #{player} has:"
-    permanentsPrintOut = []
-    permanentsPrintOut += player.keepers.map do |keeper|
-      keeper.to_s
-    end
-    permanentsPrintOut += player.creepers.map do |creeper|
-      creeper.to_s
-    end
-    debug "#{prompt}\n #{permanentsPrintOut}"
-  end
-
-  def debug(message)
-    if @debug
-      @output_stream.puts message
-    end
-  end
-
-  def trace(message)
-    if @trace
-      @output_stream.puts message
-    end
-  end
-
-  def information(message)
-    @output_stream.puts message
-  end
-
-  def log_cards(hand,prompt="Here is your current hand:")
-    debug "#{prompt}\n#{StringFormattingUtilities.indexed_display(hand)}"
-  end
-
-  def pause
-    @input_stream.gets
-  end
-end
 
 class StringFormattingUtilities
   def self.indexed_display(list)
@@ -54,15 +16,6 @@ class StringFormattingUtilities
     end
     return "#{numbering}\n#{handPrintOut}"
   end
-end
-
-class CliLogger < BaseLogger
-  def initialize(debug)
-    @output_stream = $stdout
-    @input_stream = $stdin
-    @debug = debug
-  end
-
 end
 
 class TrueCliInterface
@@ -142,13 +95,5 @@ class TestInterface
   def get_input
     input = @input_stream.gets
     input.strip
-  end
-end
-
-class TestLogger < BaseLogger
-  def initialize(input, output)
-    @input_stream = input
-    @output_stream = output
-    @debug = true
   end
 end
